feat(router): extract navigation
