+++ src/main.tsx
@@ -5,16 +5,23 @@ import { ThemeProvider } from "spartak-ui";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
 import Root from "./routes/Root";
+import Main from "./routes/Main";
 import Settings from "./routes/Settings";
 
 const router = createBrowserRouter([
   {
     path: "/",
     element: <Root />,
-  },
-  {
-    path: "settings",
-    element: <Settings />,
+    children: [
+      {
+        path: "/",
+        element: <Main />,
+      },
+      {
+        path: "settings",
+        element: <Settings />,
+      },
+    ],
   },
 ]);
